Code cleaning, icons, bug fixes for restart game logic
