+++ Croc Challenge/Model/QuestionsBox.swift
@@ -12,13 +12,7 @@ struct QuestionsBox {
     var currentTeam = 0
     var score = 0
     
-    var team = Teams.showTeam()
-//    var categoryChoise = ""
-    
-    //    var teamsNames = [Teams(name: "Ковбои", score: 0),
-    //                 Teams(name: "Стройняшки", score: 0),
-    //                 Teams(name: "Красотки", score: 0)
-    //    ]
+//    var team = Teams.showTeam()
     
     let questionsAnimals = [
         Questions(text: "Слон", answer: "Животные"),
@@ -38,7 +32,6 @@ struct QuestionsBox {
         Questions(text: "Жираф", answer: "Животные"),
         Questions(text: "Черепаха", answer: "Животные")
     ]
-    
     
     let questionsFood = [
         Questions(text: "Бургер", answer: "Еда"),
