+++ Croc Challenge/Model/TeamManager.swift
@@ -31,40 +31,22 @@ class TeamsManager {
         teamTwo = secondTeam
         ourTeam.append(teamOne!)
         ourTeam.append(teamTwo!)
+        print("team1 \(ourTeam)")
     }
-    
-  
     
     var currentTeam = 0
     
     func teamCorrectAnswer() {
         ourTeam[currentTeam].score += 1
     }
-    
-    func getCurrentTeam() -> Teams {
-        ourTeam[currentTeam]
+
+func nextTeam() -> (currentTeam: Teams, nextTeam: Teams) {
+    let current = ourTeam[currentTeam]
+    currentTeam += 1
+    if currentTeam == ourTeam.count {
+        currentTeam = 0
     }
-    
-    func nextTeam() -> (currentTeam: Teams, nextTeam: Teams) {
-        let current = ourTeam[currentTeam]
-        currentTeam += 1
-        if currentTeam == ourTeam.count {
-            currentTeam = 0
-        }
-        let next = ourTeam[currentTeam]
-        return (current, next)
-    }
-    
-//    func getScore() -> Int {
-//        return team[currentTeam].score
-//    }
-//
-   
-    
-    func rightAnswer() {
-//        if teamOne?.score < 5
-            teamOne!.score += 1
-//    } else {
-    }
-    
+    let next = ourTeam[currentTeam]
+    return (current, next)
+}
 }
