+++ Croc Challenge/Model/Teams.swift
@@ -12,11 +12,11 @@ struct Teams: Equatable {
     var score: Int
     let teamImage: UIImage
     
-    static func showTeam() -> [Teams] {
-        var post = [Teams]()
-        post.append(Teams(name: "Ковбои", score: 0, teamImage: UIImage(named: Resources.Image.cowboy)!))
-        post.append(Teams(name: "Стройняшки", score: 0, teamImage: UIImage(named: Resources.Image.cowboy)!))
-        post.append(Teams(name: "Красотки", score: 0, teamImage: UIImage(named: Resources.Image.cowboy)!))
-        return post
-    }
+//    static func showTeam() -> [Teams] {
+//        var post = [Teams]()
+//        post.append(Teams(name: "Ковбои", score: 0, teamImage: UIImage(named: Resources.Image.cowboy)!))
+//        post.append(Teams(name: "Стройняшки", score: 0, teamImage: UIImage(named: Resources.Image.cowboy)!))
+//        post.append(Teams(name: "Красотки", score: 0, teamImage: UIImage(named: Resources.Image.cowboy)!))
+//        return post
+//    }
 }
