+++ Croc Challenge/Screens/TeamViewController.swift
@@ -10,6 +10,7 @@ import UIKit
 class TeamViewController: UIViewController {
     
     let globalData = TeamsManager.shared
+    let questionNumbers = NumbersToResult.shared
     
     lazy var imageOne = createImageView(image: globalData.teamOne!.teamImage)
     lazy var labelOne = createLabelWithTeamName(title: globalData.teamOne?.name ?? "")
@@ -19,7 +20,9 @@ class TeamViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        globalData.ourTeam.removeAll()
         globalData.generateRandomTeams()
+        questionNumbers.question = 0
         self.title = "Кто играет?"
         view.backgroundColor = .clear
         setupConstraints()
